+++ test/test/datascript.cljs
@@ -745,6 +745,39 @@
                      :in   [[?monster ?heads]] ]
                   monsters)
              [[6 1 3 4 2]])))
+    
+    (testing "Min and max are using comparator instead of default compare"
+      ;; Wrong: using js '<' operator
+      ;; (apply min [:a/b :a-/b :a/c]) => :a-/b
+      ;; (apply max [:a/b :a-/b :a/c]) => :a/c
+      ;; Correct: use IComparable interface
+      ;; (sort compare [:a/b :a-/b :a/c]) => (:a/b :a/c :a-/b)
+      (is (= (d/q '[:find (min ?x) (max ?x)
+                    :in [?x ...]]
+                  [:a-/b :a/b])
+             [[:a/b :a-/b]]))
+
+      (is (= (d/q '[:find (min 2 ?x) (max 2 ?x)
+                    :in [?x ...]]
+                  [:a/b :a-/b :a/c])
+             [[[:a/b :a/c] [:a/c :a-/b]]])))
+
+    (testing "Min and max comparator order types reliably"
+      ;; Wrong: js '<' operator coerce everything to string
+      ;; (apply min ["1" :1 1]) => 1
+      ;; (apply max ["1" :1 1]) => 1
+      ;; Correct: dc/cmp-val compares types first
+      ;; (sort dc/cmp-val ["1" :1 1]) => (:1 1 "1")
+      (is (= (d/q '[:find (min ?x) (max ?x)
+                    :in [?x ...]]
+                  ["1" :1 1])
+             [[:1 "1"]]))
+
+      (is (= (d/q '[:find (min 2 ?x) (max 2 ?x)
+                    :in [?x ...]]
+                  ["1" :1 1])
+             [[[:1 1] [1 "1"]]])))
+    
 
     (testing "Grouping and parameter passing"
       (is (= (set (d/q '[ :find ?color (max ?amount ?x) (min ?amount ?x)